ATM Machine.py
This program is a simulation of a bank transaction system that allows users to perform various transactions, such as money transfers, PLN bill payments, credit purchases, cash withdrawals, and balance checks. Each transaction is protected by PIN verification to maintain security, and the user's balance will be updated after the transaction. This program also provides the option to continue the transaction or cancel and return to the main menu. -GalihDian-